TransformerMS/requirements.txt
arxiv==2.1.3
beautifulsoup4==4.13.4
datasets==3.5.0
dvclive==3.48.2
emoji==2.14.1
fastapi==0.115.12
httpx==0.28.1
joblib==1.4.2
nltk==3.9.1
pdfminer_six==20250327
pydantic==2.11.4
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
Requests==2.32.3
starlette==0.46.2
tokenizers==0.21.1
torch==2.7.0
transformers==4.49.0
uvicorn==0.34.2
